Commercial Detector 3
No code reused from other commercial detectors.  This makes the code a *lot* cleaner.  That is not to say that this wasn't influenced by them, but the code is significantly different both in organization, and in style, and in clutter.

Initially supporting Logo, Blanks, and Scene Changes.  May try to do subtitles or audio in the future.